pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "omni"
version = "0.0.1a1"
requires-python = ">=3.10"
description = "DreamLLM: Synergistic Multimodal Comprehension and Creation."
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "av",
    "tqdm",
    "fire",
    "ipdb",
    "numpy",
    "wandb",
    "isort",
    "pyzmq",
    "black",
    "boto3",
    "scipy",
    "awscli",
    "psutil",
    "dacite",
    "joblib",
    "loguru",
    "openai",
    "megfile",
    "pandas",
    "imageio",
    "isodate",
    "uvicorn",
    "fastapi",
    "pendulum",
    "requests",
    "openpyxl",
    "beartype",
    "tabulate",
    "influxdb",
    "colorama",
    "jsonlines",
    "shortuuid",
    "netifaces",
    "omegaconf",
    "smart_open",
    "ema_pytorch",
    "Levenshtein",
    "pycocotools",
    "pyinstrument",
    "editdistance",
    "nvidia-ml-py3",
    "opencv-python",
    "markdown2[all]",
    "pytorch-warmup",
    "gateloop_transformer",
    "peft==0.3.0",
    "redis>=3.4.0",
    "timm==0.6.13",
    "httpx==0.24.0",
    "einops==0.7.0",
    "gradio==3.35.2",
    "Flask>=0.10.1",
    "kornia==0.7.0",
    "msgpack>=1.0.0",
    "deepspeed==0.12.6",
    "datasets==2.16.1",
    "netifaces>=0.10.6",
    "diffusers==0.31.0",
    "hydra-core==1.3.2",
    "tokenizers==0.15.0",
    "urllib3<2,>=1.21.1",
    "pycocoevalcap==1.2",
    "einops-exts==0.0.4",
    "bitsandbytes==0.41.0",
    "transformers==4.46.1",
    "sentencepiece==0.1.99",
    "huggingface-hub==0.19.4",
    "websocket-client==0.43.0",
    "accelerate==1.6.0",    
    "vllm==0.5.0.post1",
    "xformers==0.0.26.post1",
    "vllm-flash-attn==2.5.9",
    "ms-swift==3.2.2",
    "qwen-vl-utils==0.0.10"
]
[project.scripts]
smartrun = "omni:dispatch_torchrun"

[project.urls]
Homepage = "https://dreamllm.github.io/"
Repository = "https://github.com/RunpeiDong/DreamLLM"
"Bug Tracker" = "https://github.com/RunpeiDong/DreamLLM/issues"
Changelog = "https://github.com/RunpeiDong/DreamLLM/blob/master/CHANGELOG.md"

[tool.setuptools]
packages = ["omni"]

[tool.black]
line-length = 128
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 128
